pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "fwtv"
dynamic = ["version"]
authors = [
    { name = "Leon Budnick", email = "y6q6ea9w@mail-proxy.org" },
]
description = "Verification that attendances comply with german rules. Also provide an integration to the api of FactorialHR"
readme = "README.md"
requires-python = ">=3.11, <3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU Affero General Public License v3",
    "Operating System :: OS Independent",
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "Topic :: Office/Business",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
]
keywords = [
    "FactorialHR",
    "HR",
    "Working time"
]
dependencies = [
    "factorialhr",
    "tabulate",
    "pyside6-essentials",
    "outcome",
    "trio",
]

[project.scripts]
factorial-working-time = "fwtv.main:main"

[project.urls]
"Homepage" = "https://github.com/leon1995/fwtv"
"Bug Tracker" = "https://github.com/leon1995/fwtv/issues"
[project.optional-dependencies]
test = [
    "pytest",
]
lint = [
    'black',
    'flake8',
    'mypy',
    'types-tabulate',
    'isort'
]
dev = ["fwtv[test,lint]"]

[tool.hatch.build.targets.sdist]
include = [
    "src/fwtv/**/*.py",
    "README.md",
    "pyproject.toml",
    "CHANGELOG.md"
]
support-legacy = false

[tool.hatch.build.targets.wheel]
packages = ["src/fwtv"]

[tool.hatch.version]
path = "src/fwtv/__init__.py"

[tool.pytest.ini_options]
testpaths = "tests/"

[tool.black]
line-length = 120

[tool.isort]
force_single_line = "1"
known_localfolder = "fwtv"
force_alphabetical_sort_within_sections = "1"
profile = "black"

[tool.mypy]
ignore_missing_imports = true